Handle stream errors before acting on the stream (#35)
During the setup of a Tcp/SocketStream, an error may occur which results
in self._stream_error being set.  libuv used to attempt to writing to
the stream, regardless if the stream was considered writable.

libuv/libuv@fd049399 (landed in 1.19.0, reverted in 1.19.1, and
re-landed in 1.23.0) changed uv_write to consider the writable status of
the stream.  This broke lua-client's detection of the stream open
failure, since we were only handling it when a write failure occurred.

Rather than relying on a specific stream method to fail, all stream
methods now check for self._stream_error and raise the error if it's
set.  This fixes the behavior with libuv versions that contain the
writable check, while the write method's error callback maintains the
behavior for other libuv versions.

Closes #34